@@ -19,53 +19,61 @@
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "build:client-and-server-bundles-no-prod": "ng build && ng build --app 1 --output-hashing=false",
-    "build:client-and-server-bundles": "ng build --prod && ng build --prod --app 1 --output-hashing=false",
-    "build:prerender": "npm run build:client-and-server-bundles && npm run webpack:server && npm run generate:prerender",
-    "build:ssr": "npm run build:client-and-server-bundles && npm run webpack:server",
-    "build:ssrn": "npm run build:client-and-server-bundles-no-prod && npm run webpack:server",
+    "lint": "ng lint ng-universal-demo",
+    "build:client-and-server-bundles": "ng build --prod && ng run ng-universal-demo:server:production",
+    "build:prerender": "npm run build:client-and-server-bundles && npm run compile:server && npm run generate:prerender",
+    "build:ssr": "npm run build:client-and-server-bundles && npm run compile:server",
+    "compile:server": "tsc -p server.tsconfig.json",
     "generate:prerender": "cd dist && node prerender",
-    "webpack:server": "webpack --config webpack.server.config.js --progress --colors",
     "serve:prerender": "cd dist/browser && http-server",
     "serve:ssr": "node dist/server"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/platform-server": "^5.0.0",
-    "@angular/router": "^5.0.0",
-    "@nguniversal/express-engine": "^5.0.0-beta.5",
-    "@nguniversal/module-map-ngfactory-loader": "^5.0.0-beta.5",
-    "angulartics2": "^4.5.0",
-    "core-js": "^2.4.1",
+    "@angular/animations": "^6.1.3",
+    "@angular/common": "6.1.3",
+    "@angular/compiler": "6.1.3",
+    "@angular/core": "6.1.3",
+    "@angular/forms": "6.1.3",
+    "@angular/http": "6.1.3",
+    "@angular/platform-browser": "6.1.3",
+    "@angular/platform-browser-dynamic": "6.1.3",
+    "@angular/platform-server": "^6.1.3",
+    "@angular/router": "6.1.3",
+    "@auth0/angular-jwt": "^1.2.0",
+    "@nguniversal/express-engine": "^6.0.0",
+    "@nguniversal/module-map-ngfactory-loader": "^6.0.0",
+    "angulartics2": "^6.2.0",
+    "core-js": "^2.5.7",
     "font-awesome": "^4.7.0",
     "ng2-cookies": "^1.0.12",
-    "ng2-scroll-to-el": "^1.2.1",
     "ngx-json-ld": "^0.3.1",
+    "ngx-quill": "^3.4.0",
     "pinkie-promise": "git+https://github.com/floatdrop/pinkie-promise.git",
-    "pm2": "^3.0.0",
-    "primeng": "^5.0.2",
+    "pm2": "^3.0.3",
+    "primeicons": "^1.0.0-beta.10",
+    "primeng": "^6.1.2",
+    "quill": "^1.3.6",
+    "quill-image-drop-module": "^1.0.3",
+    "quill-image-resize-module": "^3.0.0",
     "require-from-string": "https://github.com/floatdrop/require-from-string/tarball/v1.1.0",
-    "rxjs": "^5.5.2",
-    "zone.js": "^0.8.14"
+    "rxjs": "^6.2.2",
+    "rxjs-compat": "^6.2.2",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.5.0",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^5.0.0",
-    "@types/node": "^8.0.30",
+    "@angular-devkit/build-angular": "~0.7.0",
+    "@angular/cli": "6.1.4",
+    "@angular/compiler-cli": "6.1.3",
+    "@angular/language-service": "6.1.3",
+    "@types/node": "^8.10.26",
     "cpy-cli": "^1.0.1",
-    "express": "^4.15.2",
+    "express": "^4.16.3",
     "http-server": "^0.10.0",
-    "reflect-metadata": "^0.1.10",
+    "reflect-metadata": "^0.1.12",
+    "regexp-replace-loader": "^1.0.1",
     "ts-loader": "^2.3.7",
-    "typescript": "~2.4.2"
+    "typescript": "2.9.2",
+    "webpack-cli": "^3.1.0"
   }
 }
