@@ -1,8 +1,10 @@
 
+import {map, take} from 'rxjs/operators';
+
 import { Injectable } from '@angular/core';
 import { Resolve, Router, RouterStateSnapshot, ActivatedRouteSnapshot } from '@angular/router';
 
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { ObtenerEntradaResponse } from '../../dtos/response/obtenerEntradaResponse';
 import { environment } from '../../../environments/environment';
 import { EntradaService } from '../entrada.service';
@@ -22,14 +24,14 @@ export class ObtenerEntradaZonaResolverService implements Resolve<ObtenerEntrada
     const url_entrada = route.paramMap.get('url_entrada');
 
 
-    return this.entradaService.getEntrada(url_entrada).take(1).map(entrada => {
+    return this.entradaService.getEntrada(url_entrada).pipe(take(1),map(entrada => {
       if (entrada.entrada != null) {
         return entrada;
       } else { // url not found
         this.router.navigate(['/not-found']);
         return null;
       }
-    });
+    }));
   }
 
 }
