@@ -1,9 +1,11 @@
+
+import {map, take} from 'rxjs/operators';
 import { Injectable } from '@angular/core';
 import { Resolve, Router, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 import { ObtenerPaginaEtiquetaResponse } from '../../dtos/response/obtenerPaginaEtiquetaResponse';
 import { environment } from '../../../environments/environment';
 import { ClasificadorService } from '../clasificador.service';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 @Injectable()
 export class ObtenerListaEtiquetaResolverService implements Resolve<ObtenerPaginaEtiquetaResponse> {
@@ -19,23 +21,23 @@ export class ObtenerListaEtiquetaResolverService implements Resolve<ObtenerPagin
     const url = route.paramMap.get('url_etiqueta');
     const numeroPagina = route.paramMap.get('numero_pagina');
     if (numeroPagina) {
-      return this.clasificadorService.getEtiquetaPage(url, numeroPagina).take(1).map(etiqueta => {
+      return this.clasificadorService.getEtiquetaPage(url, numeroPagina).pipe(take(1),map(etiqueta => {
         if (etiqueta.etiqueta != null) {
           return etiqueta;
         } else { // url not found
           this.router.navigate(['/not-found']);
           return null;
         }
-      });
+      }),);
     } else {
-      return this.clasificadorService.getEtiqueta(url).take(1).map(etiqueta => {
+      return this.clasificadorService.getEtiqueta(url).pipe(take(1),map(etiqueta => {
         if (etiqueta.etiqueta != null) {
           return etiqueta;
         } else { // url not found
           this.router.navigate(['/not-found']);
           return null;
         }
-      });
+      }));
     }
   }
 
