@@ -3,7 +3,8 @@ import { ObtenerPaginaColeccionLibroResponse } from '../../dtos/response/obtener
 import { Resolve, Router, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 import { environment } from '../../../environments/environment';
 import { ClasificadorService } from '../clasificador.service';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
+import {map, take} from 'rxjs/operators';
 
 @Injectable()
 export class ObtenerLibroMiscelaneosResolverService implements Resolve<ObtenerPaginaColeccionLibroResponse> {
@@ -19,23 +20,27 @@ export class ObtenerLibroMiscelaneosResolverService implements Resolve<ObtenerPa
     const url = route.paramMap.get('url-libro');
     const numeroPagina = route.paramMap.get('numero_pagina');
     if (numeroPagina) {
-      return this.clasificadorService.getPaginaMiscelaneosLibroPage(url, numeroPagina).take(1).map(libroNoticias => {
-        if (libroNoticias.libro != null) {
-          return libroNoticias;
-        } else { // url not found
-          this.router.navigate(['/not-found']);
-          return null;
-        }
-      });
+      return this.clasificadorService.getPaginaMiscelaneosLibroPage(url, numeroPagina).pipe(
+        take(1),
+        map(libroNoticias => {
+          if (libroNoticias.libro != null) {
+            return libroNoticias;
+          } else { // url not found
+            this.router.navigate(['/not-found']);
+            return null;
+          }
+        }),);
     } else {
-      return this.clasificadorService.getPaginaMiscelaneosLibro(url).take(1).map(libroNoticias => {
+      return this.clasificadorService.getPaginaMiscelaneosLibro(url).pipe(
+        take(1),
+        map(libroNoticias => {
         if (libroNoticias.libro != null) {
           return libroNoticias;
         } else { // url not found
           this.router.navigate(['/not-found']);
           return null;
         }
-      });
+      }),);
     }
   }
 }
