@@ -7,8 +7,9 @@ import {
 } from '@angular/router';
 import { environment } from '../../../environments/environment';
 import { ClasificadorService } from '../clasificador.service';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { ObtenerPaginaColeccionSagaResponse } from '../../dtos/response/obtenerPaginaSagaColeccionResponse';
+import { map, take } from 'rxjs/operators';
 
 @Injectable()
 export class ObtenerSagaMiscelaneosResolverService
@@ -18,7 +19,7 @@ export class ObtenerSagaMiscelaneosResolverService
   constructor(
     private clasificadorService: ClasificadorService,
     private router: Router
-  ) { }
+  ) {}
 
   resolve(
     route: ActivatedRouteSnapshot,
@@ -30,18 +31,24 @@ export class ObtenerSagaMiscelaneosResolverService
     const url = route.paramMap.get('url-saga');
     const numeroPagina = route.paramMap.get('numero_pagina');
     if (numeroPagina) {
-      return this.clasificadorService.getPaginaMiscelaneosSagaPage(url, numeroPagina).take(1)
-        .map(miscelaneosSaga => {
-          if (miscelaneosSaga.saga != null) {
-            return miscelaneosSaga;
-          } else {
-            // url not found
-            this.router.navigate(['/not-found']);
-            return null;
-          }
-        });
+      return this.clasificadorService
+        .getPaginaMiscelaneosSagaPage(url, numeroPagina)
+        .pipe(
+          take(1),
+          map(miscelaneosSaga => {
+            if (miscelaneosSaga.saga != null) {
+              return miscelaneosSaga;
+            } else {
+              // url not found
+              this.router.navigate(['/not-found']);
+              return null;
+            }
+          })
+        );
     } else {
-      return this.clasificadorService.getPaginaMiscelaneosSaga(url).take(1).map(miscelaneosSaga => {
+      return this.clasificadorService.getPaginaMiscelaneosSaga(url).pipe(
+        take(1),
+        map(miscelaneosSaga => {
           if (miscelaneosSaga.saga != null) {
             return miscelaneosSaga;
           } else {
@@ -49,7 +56,7 @@ export class ObtenerSagaMiscelaneosResolverService
             this.router.navigate(['/not-found']);
             return null;
           }
-        });
+        }),);
     }
   }
 }
