@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import 'rxjs/add/operator/toPromise';
+
 import { environment } from '../../environments/environment';
 import { InitData } from '../dtos/initData';
 import { APP_DATA } from './app-data';
