@@ -9,6 +9,8 @@ import { UserRegistrationComponent } from './components/user-registration/user-r
 import { ForgotPasswordComponent } from './components/forgot-password/forgot-password.component';
 import { LogoutComponent } from './components/logout/logout.component';
 import { SocialLoginSuccessComponent } from './components/social-login-success/social-login-success.component';
+import { AuthGuardService } from './services/auth-guard.service';
+import { AuthService } from './services/auth.service';
 
 
 const authRoutes: Routes = [
@@ -45,6 +47,7 @@ const authRoutes: Routes = [
   ],
   exports: [
     RouterModule
-  ]
+  ],
+  providers:[AuthGuardService, AuthService]
 })
 export class AuthRoutingModule {}
