@@ -2,8 +2,7 @@ import { Component } from '@angular/core';
 
 import { NewUser, SignupStatus } from '../../dtos/login';
 import { AuthService } from '../../services/auth.service';
-import { environment } from 'environments/environment';
-
+import { environment } from '../../../../../environments/environment';
 
 @Component({
   selector: 'app-user-registration',
