@@ -2,12 +2,8 @@ import { Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-gestion',
-  template: `
-   <p>
-  		gestion Works!
-   </p>
-  `,
-  styles: []
+  templateUrl: './gestion.component.html',
+  styleUrls: ['./gestion.component.css']
 })
 export class GestionComponent implements OnInit {
 
