@@ -1,7 +1,10 @@
+
+import {throwError as observableThrowError,  Observable } from 'rxjs';
+
+import {catchError, map} from 'rxjs/operators';
 import { environment } from '../../environments/environment';
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
 import { GenrePageResponse } from '../dtos/genre/genrePageResponse';
 import { ObtenerPaginaCategoriaResponse } from '../dtos/response/obtenerPaginaCategoriaResponse';
 import { ObtenerPaginaEtiquetaResponse } from '../dtos/response/obtenerPaginaEtiquetaResponse';
@@ -36,9 +39,9 @@ export class ClasificadorService {
     if (this.log) {
       console.log("SERVER: " + this.getGenreUrl + urlGenero + '/' + numeroPagina + "/" + order);
     }
-    return this.http.get<GenrePageResponse>(this.getGenreUrl + urlGenero + '/' + numeroPagina + "/" + order)
-      .map(this.obtenerEntradaDeRespuesta)
-      .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
+    return this.http.get<GenrePageResponse>(this.getGenreUrl + urlGenero + '/' + numeroPagina + "/" + order).pipe(
+      map(this.obtenerEntradaDeRespuesta),
+      catchError((error: any) => observableThrowError(error.json().error || 'Server error')),);
   }
 
 
@@ -51,34 +54,34 @@ export class ClasificadorService {
     if (this.log) {
       console.log(urlCategoria);
     }
-    return this.http.get<ObtenerPaginaCategoriaResponse>(this.getCategoriaUrl + urlCategoria).map(this.obtenergetCategoriaDeRespuesta)
-      .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
+    return this.http.get<ObtenerPaginaCategoriaResponse>(this.getCategoriaUrl + urlCategoria).pipe(map(this.obtenergetCategoriaDeRespuesta),
+      catchError((error: any) => observableThrowError(error.json().error || 'Server error')),);
   }
 
   getCategoriaPage(urlCategoria, numeroPagina): Observable<ObtenerPaginaCategoriaResponse> {
     if (this.log) {
       console.log(urlCategoria);
     }
-    return this.http.get<ObtenerPaginaCategoriaResponse>(this.getCategoriaUrl + urlCategoria + '/' + numeroPagina)
-      .map(this.obtenergetCategoriaDeRespuesta)
-      .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
+    return this.http.get<ObtenerPaginaCategoriaResponse>(this.getCategoriaUrl + urlCategoria + '/' + numeroPagina).pipe(
+      map(this.obtenergetCategoriaDeRespuesta),
+      catchError((error: any) => observableThrowError(error.json().error || 'Server error')),);
   }
 
   getEtiqueta(urlEtiqueta): Observable<ObtenerPaginaEtiquetaResponse> {
     if (this.log) {
       console.log(urlEtiqueta);
     }
-    return this.http.get<ObtenerPaginaEtiquetaResponse>(this.getEtiquetaUrl + urlEtiqueta).map(this.obtenerEtiquetaDeRespuesta)
-      .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
+    return this.http.get<ObtenerPaginaEtiquetaResponse>(this.getEtiquetaUrl + urlEtiqueta).pipe(map(this.obtenerEtiquetaDeRespuesta),
+      catchError((error: any) => observableThrowError(error.json().error || 'Server error')),);
   }
 
   getEtiquetaPage(urlEtiqueta, numeroPagina): Observable<ObtenerPaginaEtiquetaResponse> {
     if (this.log) {
       console.log(urlEtiqueta);
     }
-    return this.http.get<ObtenerPaginaEtiquetaResponse>(this.getEtiquetaUrl + urlEtiqueta + '/' + numeroPagina)
-      .map(this.obtenerEtiquetaDeRespuesta)
-      .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
+    return this.http.get<ObtenerPaginaEtiquetaResponse>(this.getEtiquetaUrl + urlEtiqueta + '/' + numeroPagina).pipe(
+      map(this.obtenerEtiquetaDeRespuesta),
+      catchError((error: any) => observableThrowError(error.json().error || 'Server error')),);
   }
 
   
@@ -94,34 +97,34 @@ export class ClasificadorService {
     if (this.log) {
       console.log(urlLibro);
     }
-    return this.http.get<ObtenerPaginaColeccionLibroResponse>(this.getNoticiasLibroUrl + urlLibro).map(this.obtenerColeccionLibroDeRespuesta)
-      .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
+    return this.http.get<ObtenerPaginaColeccionLibroResponse>(this.getNoticiasLibroUrl + urlLibro).pipe(map(this.obtenerColeccionLibroDeRespuesta),
+      catchError((error: any) => observableThrowError(error.json().error || 'Server error')),);
   }
 
   getPaginaNoticiasLibroPage(urlLibro, numeroPagina): Observable<ObtenerPaginaColeccionLibroResponse> {
     if (this.log) {
       console.log(urlLibro);
     }
-    return this.http.get<ObtenerPaginaColeccionLibroResponse>(this.getNoticiasLibroUrl + urlLibro + '/' + numeroPagina)
-      .map(this.obtenerColeccionLibroDeRespuesta)
-      .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
+    return this.http.get<ObtenerPaginaColeccionLibroResponse>(this.getNoticiasLibroUrl + urlLibro + '/' + numeroPagina).pipe(
+      map(this.obtenerColeccionLibroDeRespuesta),
+      catchError((error: any) => observableThrowError(error.json().error || 'Server error')),);
   }
 
   getPaginaMiscelaneosLibro(urlLibro): Observable<ObtenerPaginaColeccionLibroResponse> {
     if (this.log) {
       console.log(urlLibro);
     }
-    return this.http.get<ObtenerPaginaColeccionLibroResponse>(this.getMiscelaneosLibroUrl + urlLibro).map(this.obtenerColeccionLibroDeRespuesta)
-      .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
+    return this.http.get<ObtenerPaginaColeccionLibroResponse>(this.getMiscelaneosLibroUrl + urlLibro).pipe(map(this.obtenerColeccionLibroDeRespuesta),
+      catchError((error: any) => observableThrowError(error.json().error || 'Server error')),);
   }
 
   getPaginaMiscelaneosLibroPage(urlLibro, numeroPagina): Observable<ObtenerPaginaColeccionLibroResponse> {
     if (this.log) {
       console.log(urlLibro);
     }
-    return this.http.get<ObtenerPaginaColeccionLibroResponse>(this.getMiscelaneosLibroUrl + urlLibro + '/' + numeroPagina)
-      .map(this.obtenerColeccionLibroDeRespuesta)
-      .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
+    return this.http.get<ObtenerPaginaColeccionLibroResponse>(this.getMiscelaneosLibroUrl + urlLibro + '/' + numeroPagina).pipe(
+      map(this.obtenerColeccionLibroDeRespuesta),
+      catchError((error: any) => observableThrowError(error.json().error || 'Server error')),);
   }
 
   private obtenerColeccionLibroDeRespuesta(res: ObtenerPaginaColeccionLibroResponse) {
@@ -132,26 +135,30 @@ export class ClasificadorService {
     if (this.log) {
       console.log(parametrosBusqueda);
     }
-    return this.http.get<ObtenerPaginaColeccionLibroResponse>(this.getBusquedaUrl + parametrosBusqueda)
-      .map(this.obtenerColeccionLibroDeRespuesta)
-      .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
+    return this.http.get<ObtenerPaginaColeccionLibroResponse>(this.getBusquedaUrl + parametrosBusqueda).pipe(
+      map(this.obtenerPaginaBusquedaResponse),
+      catchError((error: any) => observableThrowError(error.json().error || 'Server error')),);
+  }
+
+  private obtenerPaginaBusquedaResponse(res: ObtenerPaginaBusquedaResponse) {
+    return res;
   }
 
   getEditor(urlEditor): Observable<ObtenerPaginaRedactorResponse> {
     if (this.log) {
       console.log(urlEditor);
     }
-    return this.http.get<ObtenerPaginaRedactorResponse>(this.getEditorUrl + urlEditor).map(this.obtenerEditorDeRespuesta)
-      .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
+    return this.http.get<ObtenerPaginaRedactorResponse>(this.getEditorUrl + urlEditor).pipe(map(this.obtenerEditorDeRespuesta),
+      catchError((error: any) => observableThrowError(error.json().error || 'Server error')),);
   }
 
   getEditorPage(urlEditor, numeroPagina): Observable<ObtenerPaginaRedactorResponse> {
     if (this.log) {
       console.log(urlEditor);
     }
-    return this.http.get<ObtenerPaginaRedactorResponse>(this.getEditorUrl + urlEditor + '/' + numeroPagina)
-      .map(this.obtenerEditorDeRespuesta)
-      .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
+    return this.http.get<ObtenerPaginaRedactorResponse>(this.getEditorUrl + urlEditor + '/' + numeroPagina).pipe(
+      map(this.obtenerEditorDeRespuesta),
+      catchError((error: any) => observableThrowError(error.json().error || 'Server error')),);
   }
 
   private obtenerEditorDeRespuesta(res: ObtenerPaginaRedactorResponse) {
@@ -162,17 +169,17 @@ export class ClasificadorService {
     if (this.log) {
       console.log(urlEditorial);
     }
-    return this.http.get<ObtenerPaginaEditorialResponse>(this.getEditorialUrl + urlEditorial).map(this.obtenerEditorialDeRespuesta)
-      .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
+    return this.http.get<ObtenerPaginaEditorialResponse>(this.getEditorialUrl + urlEditorial).pipe(map(this.obtenerEditorialDeRespuesta),
+      catchError((error: any) => observableThrowError(error.json().error || 'Server error')),);
   }
 
   getEditorialPage(urlEditorial, numeroPagina): Observable<ObtenerPaginaEditorialResponse> {
     if (this.log) {
       console.log(urlEditorial);
     }
-    return this.http.get<ObtenerPaginaEditorialResponse>(this.getEditorialUrl + urlEditorial + '/' + numeroPagina)
-      .map(this.obtenerEditorialDeRespuesta)
-      .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
+    return this.http.get<ObtenerPaginaEditorialResponse>(this.getEditorialUrl + urlEditorial + '/' + numeroPagina).pipe(
+      map(this.obtenerEditorialDeRespuesta),
+      catchError((error: any) => observableThrowError(error.json().error || 'Server error')),);
   }
 
   private obtenerEditorialDeRespuesta(res: ObtenerPaginaEditorialResponse) {
@@ -183,34 +190,34 @@ export class ClasificadorService {
     if (this.log) {
       console.log(urlEditorial);
     }
-    return this.http.get<ObtenerPaginaColeccionSagaResponse>(this.getNoticiasSagaUrl + urlEditorial).map(this.obtenerObtenerPaginaSagaColeccionResponse)
-      .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
+    return this.http.get<ObtenerPaginaColeccionSagaResponse>(this.getNoticiasSagaUrl + urlEditorial).pipe(map(this.obtenerObtenerPaginaSagaColeccionResponse),
+      catchError((error: any) => observableThrowError(error.json().error || 'Server error')),);
   }
 
   getPaginaNoticiasSagaPage(urlEditorial, numeroPagina): Observable<ObtenerPaginaColeccionSagaResponse> {
     if (this.log) {
       console.log(urlEditorial);
     }
-    return this.http.get<ObtenerPaginaColeccionSagaResponse>(this.getNoticiasSagaUrl + urlEditorial + '/' + numeroPagina)
-      .map(this.obtenerObtenerPaginaSagaColeccionResponse)
-      .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
+    return this.http.get<ObtenerPaginaColeccionSagaResponse>(this.getNoticiasSagaUrl + urlEditorial + '/' + numeroPagina).pipe(
+      map(this.obtenerObtenerPaginaSagaColeccionResponse),
+      catchError((error: any) => observableThrowError(error.json().error || 'Server error')),);
   }
 
   getPaginaMiscelaneosSaga(urlEditorial): Observable<ObtenerPaginaColeccionSagaResponse> {
     if (this.log) {
       console.log(urlEditorial);
     }
-    return this.http.get<ObtenerPaginaColeccionSagaResponse>(this.getMiscelaneosSagaUrl + urlEditorial).map(this.obtenerObtenerPaginaSagaColeccionResponse)
-      .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
+    return this.http.get<ObtenerPaginaColeccionSagaResponse>(this.getMiscelaneosSagaUrl + urlEditorial).pipe(map(this.obtenerObtenerPaginaSagaColeccionResponse),
+      catchError((error: any) => observableThrowError(error.json().error || 'Server error')),);
   }
 
   getPaginaMiscelaneosSagaPage(urlEditorial, numeroPagina): Observable<ObtenerPaginaColeccionSagaResponse> {
     if (this.log) {
       console.log(urlEditorial);
     }
-    return this.http.get<ObtenerPaginaColeccionSagaResponse>(this.getMiscelaneosSagaUrl + urlEditorial + '/' + numeroPagina)
-      .map(this.obtenerObtenerPaginaSagaColeccionResponse)
-      .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
+    return this.http.get<ObtenerPaginaColeccionSagaResponse>(this.getMiscelaneosSagaUrl + urlEditorial + '/' + numeroPagina).pipe(
+      map(this.obtenerObtenerPaginaSagaColeccionResponse),
+      catchError((error: any) => observableThrowError(error.json().error || 'Server error')),);
   }
 
   private obtenerObtenerPaginaSagaColeccionResponse(res: ObtenerPaginaColeccionSagaResponse) {
