@@ -1,29 +1,31 @@
+import { throwError as observableThrowError, Observable } from 'rxjs';
 import { Injectable } from '@angular/core';
 import { environment } from '../../environments/environment';
 import { HttpClient } from '@angular/common/http';
 import { ComentarioRequest } from '../dtos/request/comentarioRequest';
-import { Observable } from 'rxjs/Observable';
 import { GuardarComentarioResponse } from '../dtos/response/guardarComentarioResponse';
+import { catchError, map } from 'rxjs/operators';
 
 @Injectable()
 export class ComentariosService {
-
   private log = environment.log;
 
   private addComentarioUrl = environment.addComentarioUrl;
 
-  constructor(private http: HttpClient) { }
+  constructor(private http: HttpClient) {}
 
-  guardarComentario(comentarioRequest: ComentarioRequest): Observable<GuardarComentarioResponse> {
-
+  guardarComentario(
+    comentarioRequest: ComentarioRequest
+  ): Observable<GuardarComentarioResponse> {
     return this.http
       .post<GuardarComentarioResponse>(this.addComentarioUrl, comentarioRequest)
-      .map(this.extractGuardarComentarioResponse)
-      .catch(error => Observable.throw(error || 'Server error'));
+      .pipe(
+        map(this.extractGuardarComentarioResponse),
+        catchError(error => observableThrowError(error || 'Server error'))
+      );
   }
 
   private extractGuardarComentarioResponse(res: GuardarComentarioResponse) {
     return res;
   }
-
 }
