@@ -1,12 +1,15 @@
+
+import {throwError as observableThrowError,  Observable } from 'rxjs';
+
+import {catchError, map} from 'rxjs/operators';
 import { Injectable } from '@angular/core';
 import { environment } from '../../environments/environment';
 import { ObtenerIndexDataResponse } from '../dtos/response/obtenerIndexDataResponse';
 
 import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/take';
+
+
+
 import { Cookie } from 'ng2-cookies';
 
 @Injectable()
@@ -25,14 +28,14 @@ export class IndexDataService {
       console.log('getIndexData()');
     }
 
-    this.resultados = this.http.get(this.indexDataUrl)
-      .map(this.extractIndexData)
-      .catch(error => Observable.throw(error || 'Server error').share());
+    this.resultados = this.http.get(this.indexDataUrl).pipe(
+      map(this.extractIndexData),
+      catchError(error => observableThrowError(error || 'Server error')),);
     return this.resultados;
 
   }
 
-  private extractIndexData(res: Response) {
+  private extractIndexData(res: ObtenerIndexDataResponse) {
     return res;
   }
 
