@@ -1,18 +1,18 @@
+import { throwError as observableThrowError, Observable } from 'rxjs';
 import { Injectable } from '@angular/core';
 import { environment } from '../../environments/environment';
 import { Libro } from '../dtos/libro';
 import { Genero } from '../dtos/genre/genero';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Cookie } from 'ng2-cookies';
-import { Observable } from 'rxjs/Observable';
 import { FichaLibro } from '../dtos/fichaLibro';
 import { GuardarLibroResponse } from '../dtos/response/guardarLibroResponse';
 import { GuardarGeneroResponse } from '../dtos/genre/guardarGeneroResponse';
 import { JsonAdapterService } from './util/json-adapter.service';
+import { catchError, map } from 'rxjs/operators';
 
 @Injectable()
 export class LibroService {
-
   private log = environment.log;
 
   private librosUrl = environment.librosUrl;
@@ -21,38 +21,44 @@ export class LibroService {
   private addGeneroUrl = environment.addGeneroUrl;
   private obtenerLibroUrl = environment.obtenerLibroUrl;
 
-
   results: string[];
 
   librosList: Libro[] = new Array();
   allGenerosList: Genero[] = new Array();
 
-  constructor(private http: HttpClient, private jsonAdapter: JsonAdapterService) { }
+  constructor(
+    private http: HttpClient,
+    private jsonAdapter: JsonAdapterService
+  ) {}
 
   getLibros(): Promise<Libro[]> {
     if (this.log) {
       console.log('access_token: ' + Cookie.get('access_token'));
-      console.log('_ga: '+ Cookie.get('_ga'))
+      console.log('_ga: ' + Cookie.get('_ga'));
     }
     const headers = new HttpHeaders({
       'Content-type': 'application/json',
-      'Authorization': Cookie.get('access_token')
+      Authorization: Cookie.get('access_token')
     });
     this.librosList = new Array();
-    return this.http.get(this.librosUrl, { headers: headers }).toPromise().then((resp: Response) => {
-      for (const numLibro of Object.keys(resp)) {
-        const l = this.jsonAdapter.adaptarLibro(resp[numLibro])
-        this.librosList.push(l);
-      }
+    return this.http
+      .get(this.librosUrl, { headers: headers })
+      .toPromise()
+      .then((resp: Response) => {
+        for (const numLibro of Object.keys(resp)) {
+          const l = this.jsonAdapter.adaptarLibro(resp[numLibro]);
+          this.librosList.push(l);
+        }
 
-      return this.librosList;
-
-    });
+        return this.librosList;
+      });
   }
 
   getLibro(urlLibro: string): Observable<FichaLibro> {
-    return this.http.get<FichaLibro>(this.obtenerLibroUrl + urlLibro).map(this.extractLibro)
-      .catch(error => Observable.throw(error || 'Server error'));
+    return this.http
+      .get<FichaLibro>(this.obtenerLibroUrl + urlLibro).pipe(
+      map(this.extractLibro),
+      catchError(error => observableThrowError(error || 'Server error')),);
   }
 
   private extractLibro(res: FichaLibro) {
@@ -62,12 +68,14 @@ export class LibroService {
   guardarLibro(libro: Libro): Observable<GuardarLibroResponse> {
     const headers = new HttpHeaders({
       'Content-type': 'application/json',
-      'Authorization': Cookie.get('access_token')
+      Authorization: Cookie.get('access_token')
     });
     return this.http
-      .post<GuardarLibroResponse>(this.addLibroUrl, JSON.stringify(libro), { headers: headers })
-      .map(this.extractGuardarLibroResponse)
-      .catch(error => Observable.throw(error || 'Server error'));
+      .post<GuardarLibroResponse>(this.addLibroUrl, JSON.stringify(libro), {
+        headers: headers
+      }).pipe(
+      map(this.extractGuardarLibroResponse),
+      catchError(error => observableThrowError(error || 'Server error')),);
   }
 
   private extractGuardarLibroResponse(res: GuardarLibroResponse) {
@@ -77,12 +85,12 @@ export class LibroService {
   guardarGenero(genero: Genero): Observable<GuardarGeneroResponse> {
     const headers = new HttpHeaders({
       'Content-type': 'application/json',
-      'Authorization': Cookie.get('access_token')
+      Authorization: Cookie.get('access_token')
     });
     return this.http
-      .post(this.addGeneroUrl, JSON.stringify(genero), { headers: headers })
-      .map(this.extractGuardarGeneroResponse)
-      .catch(error => Observable.throw(error || 'Server error'));
+      .post(this.addGeneroUrl, JSON.stringify(genero), { headers: headers }).pipe(
+      map(this.extractGuardarGeneroResponse),
+      catchError(error => observableThrowError(error || 'Server error')),);
   }
 
   private extractGuardarGeneroResponse(res: GuardarGeneroResponse) {
@@ -92,10 +100,12 @@ export class LibroService {
   getGeneros(): Observable<Genero[]> {
     const headers = new HttpHeaders({
       'Content-type': 'application/json',
-      'Authorization': Cookie.get('access_token')
+      Authorization: Cookie.get('access_token')
     });
-    return this.http.get<Genero[]>(this.generosUrl, { headers: headers }).map(this.extractGeneros)
-      .catch(error => Observable.throw(error || 'Server error'));
+    return this.http.get<Genero[]>(this.generosUrl, { headers: headers }).pipe(
+      map(this.extractGeneros),
+      catchError(error => observableThrowError(error || 'Server error'))
+    );
   }
 
   private extractGeneros(res: Genero[]) {
@@ -107,4 +117,3 @@ export class LibroService {
     return Promise.reject(error.message || error);
   }
 }
-
