@@ -1,27 +1,36 @@
-// Work around for https://github.com/angular/angular-cli/issues/7200
-
 const path = require('path');
 const webpack = require('webpack');
 
 module.exports = {
+  mode: 'none',
   entry: {
-    // This is our Express server for Dynamic universal
     server: './server.ts',
-    // This is an example of Static prerendering (generative)
-    prerender: './prerender.ts'
   },
   target: 'node',
-  resolve: { extensions: ['.ts', '.js'] },
-  // Make sure we include all node_modules etc
-  externals: [/(node_modules|main\..*\.js)/,],
+  resolve: {
+    extensions: ['.ts', '.js']
+  },
+  optimization: {
+    minimize: false
+  },
   output: {
     // Puts the output at the root of the dist folder
     path: path.join(__dirname, 'dist'),
     filename: '[name].js'
   },
   module: {
-    rules: [
-      { test: /\.ts$/, loader: 'ts-loader' }
+    rules: [{
+        test: /\.ts$/,
+        loader: 'ts-loader'
+      },
+      {
+        // Mark files inside `@angular/core` as using SystemJS style dynamic imports.
+        // Removing this will cause deprecation warnings to appear.
+        test: /(\\|\/)@angular(\\|\/)core(\\|\/).+\.js$/,
+        parser: {
+          system: true
+        },
+      },
     ]
   },
   plugins: [
@@ -34,9 +43,7 @@ module.exports = {
     new webpack.ContextReplacementPlugin(
       // fixes WARNING Critical dependency: the request of a dependency is an expression
       /(.+)?express(\\|\/)(.+)?/,
-      path.join(__dirname, 'src'),
-      {}
+      path.join(__dirname, 'src'), {}
     )
   ]
 }
-  
